Update CLOUD_HOWTO.md
Fixed errors that occured to me while trying to build with BKCL

I actually have no idea if this will change the way how the plugin will be built. But replacing compileOnlyApi > CompileOnly and replacing the val > def function in the shadowJar task will make it so it will not give problems.